FlatDHCPManager: Pass admin cxtx to update_dhcp.
Updates the Nova FlatDHCPManager so it passes an admin
context to update_dhcp. In some cases (linux_net) the
network driver may make use of DB calls which require an
admin context.

This fixes an issue where resize operations (confirm, revert)
would fail if the user was not an admin.

This is fallout from 14e2ed3 where we stopped passing admin
contexts for resize across RPC.

Change-Id: I718c40f0df0b2955626033cf3b6f78323527e9cf